Add JsonPositionedTextReaderTest for non seekable
